Working from baseline revision 85
01 Mar 10

Added GameSettings.GameName and the ability to set this from the NewGame wizard. The GameName is used to name the ServerState and GameSettings files as GameName.state and GameName.settings.

Fixed an issue (pointer duplication) that caused the console to loose track of the current turn between sessions.

Fixed an issue that was causing the console to load all race files, rather than just the races in the current game.

Nova GUI was not displaying the star map. Copied old Main function back into Nova GUI to fix and commented out Main function in program.cs.

03 Mar 10

More work on getting the GameName to be used as the Game directory.
Allowed the race files to be stored in a separate directory to the Game Files. Moved the current races to a sub directory of the Game Files.

Intel split up per race (still contains 'everything': needs to be filtered per race).

04 Mar 10

Moved some of the Turn/Intel loading code from ClientState to IntelReader.
Fixed an issue with RaceComponents that caused all available components to disapear after first turn.

Still chasing why ship designs disappear between turns...

06 Mar 10

Merging of changes with revision 89. Require adding back in some defensive code on file opening to check the file exists.



git-svn-id: svn://svn.code.sf.net/p/stars-nova/svn/trunk@90 03d337be-0f81-4a3b-8898-4a24ff344f0f